# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Progress API",
    "description": "GitHub multi-repo project tracking tool.",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/reports": {
      "get": {
        "tags": ["reports"],
        "summary": "List Reports",
        "description": "Paginated list of report summaries (newest first).",
        "operationId": "list_reports_api_v1_reports_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "report_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Report Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ReportSummary_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}": {
      "get": {
        "tags": ["reports"],
        "summary": "Get Report",
        "description": "Single report detail with rendered HTML (spec 12).",
        "operationId": "get_report_api_v1_reports__report_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/{report_id}/raw": {
      "get": {
        "tags": ["reports"],
        "summary": "Get Report Raw",
        "description": "Raw markdown body for the report (spec 12).",
        "operationId": "get_report_raw_api_v1_reports__report_id__raw_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RawMarkdownResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_ReportSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReportSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "default": 1
          },
          "page_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "default": 20
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total",
            "default": 0
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["items"],
        "title": "PaginatedResponse[ReportSummary]"
      },
      "RawMarkdownResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "markdown": {
            "type": "string",
            "title": "Markdown"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "report_type", "title", "markdown"],
        "title": "RawMarkdownResponse",
        "description": "Raw markdown response for ``GET /api/v1/reports/{id}/raw``."
      },
...
      "ReportDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "commit_hash": {
            "type": "string",
            "title": "Commit Hash"
          },
          "previous_commit_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Commit Hash"
          },
          "commit_count": {
            "type": "integer",
            "title": "Commit Count"
          },
          "markpost_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markpost Url"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "rendered_html": {
            "type": "string",
            "title": "Rendered Html"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "report_type",
          "title",
          "commit_hash",
          "commit_count",
          "content",
          "rendered_html",
          "created_at"
        ],
        "title": "ReportDetail",
        "description": "Full report including rendered HTML content (spec 12)."
      },
      "ReportSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "report_type": {
            "type": "string",
            "title": "Report Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "commit_hash": {
            "type": "string",
            "title": "Commit Hash"
          },
          "previous_commit_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Commit Hash"
          },
          "commit_count": {
            "type": "integer",
            "title": "Commit Count"
          },
          "markpost_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markpost Url"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": ["id", "report_type", "title", "commit_hash", "commit_count", "created_at"],
        "title": "ReportSummary",
        "description": "Lightweight report representation for list responses."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}